Remove unnecessary factory
