+++ core/src/main/java/com/novoda/noplayer/subtitles/SystemCaptionPreferences.java
@@ -1,7 +1,13 @@
 package com.novoda.noplayer.subtitles;
 
 import android.content.Context;
+import android.os.Build;
 import android.support.annotation.NonNull;
+import android.support.annotation.RequiresApi;
+import android.view.accessibility.CaptioningManager;
+
+import static android.os.Build.VERSION_CODES.KITKAT;
+import static java.util.Objects.requireNonNull;
 
 public class SystemCaptionPreferences {
 
@@ -13,13 +19,40 @@ public class SystemCaptionPreferences {
 
     @NonNull
     public SubtitlesStyle getStyle() {
-        return getStyle(context);
+        if (Build.VERSION.SDK_INT >= KITKAT) {
+            final CaptioningManager captioningManager = requireCaptionManager();
+            CaptioningManager.CaptionStyle systemStyle = captioningManager.getUserStyle();
+            float fontScale = captioningManager.getFontScale();
+            return new KitkatSubtitlesStyle(systemStyle, fontScale);
+        } else {
+            return NO_STYLE;
+        }
     }
 
     @NonNull
-    private static SubtitlesStyle getStyle(Context context) {
-        return SubtitlesStyleFactory.create(context);
+    @RequiresApi(api = KITKAT)
+    private CaptioningManager requireCaptionManager() {
+        Object service = context.getSystemService(Context.CAPTIONING_SERVICE);
+        return requireNonNull((CaptioningManager) service);
     }
 
+    private static final SubtitlesStyle NO_STYLE = new SubtitlesStyle() {
+
+        @Override
+        public int backgroundColorOr(int fallbackColor) {
+            return fallbackColor;
+        }
+
+        @Override
+        public int foregroundColorOr(int fallbackColor) {
+            return fallbackColor;
+        }
+
+        @Override
+        public float scaleTextSize(float textSize) {
+            return textSize;
+        }
+
+    };
 
 }
